requirements.txt
nextcord==2.0.0a3
python-dotenv==0.15.0
mysql-connector-python==8.0.12
